fix(android): move log zip file outside of log folder (#3677)
Fixes #3330

The when the log zip file is created, it's created as an empty file in
the `logs` directory. The `logs` directory is then zipped, and written
to the zip file. The empty file remains as an artifact in the final zip
file.

This moves the location of the zip file outside of the `logs` directory,
where it will be cleaned up later.