fix(connlib): set socket to IPv6 only before binding address (#3701)
Binding the IPv6 socket would always fail because we do it _after_ the
IPv4 socket. However, without setting the socket to `IP6_ONLY`, binding
the unspecified IPv6 address (`::`) also wants to bind to IPv4
(depending on the kernel settings). To avoid this, we need to first
configure the socket properly and then bind it to the given address.

Co-authored-by: Jamil <jamilbk@users.noreply.github.com>