feat(snownet): optimise which channels we bind (#3709)
Currently, we bind a lot of TURN channels on our relays because we bind
a channel to each candidate on each relay. With every node having
usually 4 relays, that results in 16 channels per connection just for
the relay candidates.

We can bring this down optimistically by first checking if the remote's
candidate is a relay candidate and happens to be on a relay that we are
also using. In that case, we only bind the channel on that one.

That should also improve latency when data needs to be relayed because
we reduce the number of hops by 1 and don't send traffic between two
relays.

Additionally, there is no reason to bind channels for host candidates.