fix(gateway): expire resources (#3722)
I forgot to actually call the expire resources function after the
refactor 🤦

This will be much cleaned up in a PR that I'm working on to eliminate
the `peers_by_id`/`peers_by_ip` maps.

In the mean time let's merge this asap since the gateway not expiring
resources is a security hole.