chore(snownet): reduce log levels (#3777)
Both of these happen quite often and aren't really of concern for
day-to-day operation. Binding a new channel is still a state change but
after using the clients for a bit, it seems that it is not an important
enough state change to actually tell the user about (which I assume will
likely log on `INFO`).

Similarly, dropping a packet because we don't have a channel happens
more often now because we've optimised, which addresses we bind channels
to.