feat(snownet): reduce connection setup latency (#3772)
Previously, we pretty much always lost the wireguard handshake packet,
causing us to wait for the rekey-timeout before we try again.

We can fix this by first checking that we actually have a socket that we
can send the encapsulated packet on. Additionally, we can directly force
a wireguard handshake as soon as we discover the first socket to the
remote.

This reduces the setup latency to ~3 seconds in my testing.

Resolves: #3779.