refactor(connlib): replace intent timer with explicit throttling (#3778)
The reference that is specified as part of the connection intent
fulfills one particular purpose: To avoid accepting connection details
for a "stale" intent, i.e. a previous one that we sent for the same
resource.

With the move to `phoenix-channel` in #3682, we can no longer specify
the reference explicitly. Instead, sending a message to the portal gives
us an `OutboundRequestId`.

To make the transition in #3682 easier, we emulate this behaviour here
temporarily in the `ControlPlane` of the clients.