refactor(connlib): fail tunnel in case the device gets closed (#3981)
Currently, we have some duplication between client and gateway in how we
handle a closed TUN device (i.e. reading a 0-length packet). Previously,
we would just deactivate the tunnel as we can't do anything when the
device is gone. I don't think that is actually helpful. Instead, we
should fail the tunnel if the device gets closed, similar to any other
IO error on the device.\

This reduces duplication between the two eventloops.